MatrixPilot_FreeRTOS: initial pass at updating with my latest working copy. more to come - not intending to document individual changes with the branch. DO NOT USE
git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPilot_FreeRTOS@3166 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ libDCM/estAltitude.h
@@ -23,13 +23,13 @@
 #define ESTALTITUDE_H
 
 
-void udb_barometer_callback(long pressure, int temperature, char status);
+void udb_barometer_callback(long pressure, int16_t temperature, char status);
 void altimeter_calibrate(void);
 void estAltitude(void);
 
 inline long get_barometer_altitude(void);
 inline long get_barometer_pressure(void);
-inline int get_barometer_temperature(void);
+inline int16_t get_barometer_temperature(void);
 inline long get_barometer_agl_altitude(void);
